+++ cpp/src/io/csv/csv_reader.cu
@@ -66,7 +66,9 @@ using cu_reccnt_t = unsigned long long int;
 using cu_recstart_t = unsigned long long int;
 
 
-//-- define the structure for raw data handling - for internal use
+/**---------------------------------------------------------------------------*
+ * @brief Struct used for internal parsing state
+ *---------------------------------------------------------------------------**/
 typedef struct raw_csv_ {
     char *				data;			// on-device: the raw unprocessed CSV data - loaded as a large char * array
     cu_recstart_t*		recStart;		// on-device: Starting position of the records.
@@ -83,14 +85,16 @@ typedef struct raw_csv_ {
     vector<string>		col_names;		// host: array of column names
     bool* 				h_parseCol;		// host   : array of booleans stating if column should be parsed in reading process: parseCol[x]=false means that the column x needs to be filtered out.
     bool* 				d_parseCol;		// device : array of booleans stating if column should be parsed in reading process: parseCol[x]=false means that the column x needs to be filtered out.
-    long 				header_row;		// Row id of the header
 
     long        byte_range_offset;  // offset into the data to start parsing
     long        byte_range_size;    // length of the data of interest to parse
 
-	gdf_size_type  		nrows;       	// number of rows of file to read. default is -1, and all rows are read in this case
-	gdf_size_type		skiprows;       // number of rows at the start of the file to skip, default is 0
-	gdf_size_type		skipfooter;     // number of rows at the bottom of the file to skip, default is 0
+    gdf_size_type header_row;       ///< host: Row index of the header
+    gdf_size_type nrows;            ///< host: Number of rows to read. -1 for all rows
+    gdf_size_type skiprows;         ///< host: Number of rows to skip from the start
+    gdf_size_type skipfooter;       ///< host: Number of rows to skip from the end
+    std::vector<char> header;       ///< host: Header row data, for parsing column names
+    string prefix;                  ///< host: Prepended to column ID if there is no header or input column names
 
     rmm::device_vector<int32_t>	d_trueValues;		// device: array of values to recognize as true
     rmm::device_vector<int32_t>	d_falseValues;		// device: array of values to recognize as false
@@ -194,6 +198,85 @@ __device__ void setBit(gdf_valid_type* address, int bit) {
 		return base_padding + num_columns * column_bytes; 
 	}
 }
+/**
+* @brief Removes the first and Last quote in the string
+*/
+string removeQuotes(string str, char quotechar) {
+	// Exclude first and last quotation char
+	const size_t first_quote = str.find(quotechar);
+	if (first_quote != string::npos) {
+		str.erase(first_quote, 1);
+	}
+	const size_t  last_quote = str.rfind(quotechar);
+	if (last_quote != string::npos) {
+		str.erase(last_quote, 1);
+	}
+
+	return str;
+}
+
+/**
+ * @brief Parse the first row to set the column names in the raw_csv parameter 
+ *
+ * The first row can be either the header row, or the first data row
+ *
+ * @param[in,out] raw_csv Structure containing the csv parsing parameters
+ * and intermediate results
+ * 
+ * @return gdf_error with error code on failure, otherwise GDF_SUCCESS
+*/
+gdf_error setColumnNamesFromCsv(raw_csv_t* raw_csv) {
+	vector<char> first_row = raw_csv->header;
+	// No header, read the first data row
+	if (first_row.empty()) {
+		cu_recstart_t first_row_len{};
+		// If file only contains one row, raw_csv->recStart[1] is not valid
+		if (raw_csv->num_records > 1) {
+			CUDA_TRY(cudaMemcpy(&first_row_len, raw_csv->recStart + 1, sizeof(cu_recstart_t), cudaMemcpyDefault));
+		}
+		else {
+			// File has one row - use the file size for the row size
+			first_row_len = raw_csv->num_bytes / sizeof(char);
+		}
+		first_row.resize(first_row_len);
+		CUDA_TRY(cudaMemcpy(first_row.data(), raw_csv->data, raw_csv->num_bytes, cudaMemcpyDefault));
+	}
+
+	int num_cols = 0;
+
+	bool quotation	= false;
+	for (size_t pos = 0, prev = 0; pos < first_row.size(); ++pos) {
+		// Flip the quotation flag if current character is a quotechar
+		if(first_row[pos] == raw_csv->opts.quotechar) {
+			quotation = !quotation;
+		}
+		else if (!quotation &&
+				 (first_row[pos] == raw_csv->opts.delimiter ||
+				 first_row[pos] == raw_csv->opts.terminator)) {
+			// Got to the end of a column
+			if (raw_csv->header_row >= 0) {
+				// first_row is the header, add the column name
+				string new_col_name(first_row.data() + prev, pos - prev);
+				raw_csv->col_names.push_back(removeQuotes(new_col_name, raw_csv->opts.quotechar));
+			}
+			else {
+				// first_row is the first data row, add the automatically generated name
+				raw_csv->col_names.push_back(raw_csv->prefix + std::to_string(num_cols));
+			}
+			num_cols++;
+
+			// Skip adjacent delimiters if delim_whitespace is set
+			while (raw_csv->opts.multi_delimiter &&
+				   pos < first_row.size() &&
+				   first_row[pos] == raw_csv->opts.delimiter && 
+				   first_row[pos + 1] == raw_csv->opts.delimiter) {
+				++pos;
+			}
+			prev = pos + 1;
+		}
+	}
+	return GDF_SUCCESS;
+}
 
 /**---------------------------------------------------------------------------*
  * @brief Read in a CSV file, extract all fields and return 
@@ -221,11 +304,14 @@ gdf_error read_csv(csv_read_arg *args)
 	raw_csv->skiprows = args->skiprows;
 	raw_csv->skipfooter = args->skipfooter;
 	raw_csv->nrows = args->nrows;
+	raw_csv->prefix = args->prefix == nullptr ? "" : string(args->prefix);
 
 	if (args->delim_whitespace) {
 		raw_csv->opts.delimiter = ' ';
+		raw_csv->opts.multi_delimiter = true;
 	} else {
 		raw_csv->opts.delimiter = args->delimiter;
+		raw_csv->opts.multi_delimiter = false;
 	}
 	if (args->windowslinetermination) {
 		raw_csv->opts.terminator = '\n';
@@ -434,43 +520,25 @@ gdf_error read_csv(csv_read_arg *args)
 
 	// Check if the user gave us a list of column names
 	if(args->names == nullptr) {
-		int h_num_cols = 0;
-		// Getting the first row of data from the file. We will parse the data to find lineterminator as
-		// well as the column delimiter.
-		cu_recstart_t second_rec_start;
-		CUDA_TRY(cudaMemcpy(&second_rec_start, raw_csv->recStart + 1, sizeof(cu_recstart_t), cudaMemcpyDefault));
-		vector<char> first_row(second_rec_start);
-		CUDA_TRY(cudaMemcpy(first_row.data(), raw_csv->data, sizeof(char) * first_row.size(), cudaMemcpyDefault));
 
-		if (raw_csv->header_row >= 0) {
-			// Parse header row to assign column names; does not currently handle
-			// quotations to avoid added parsing complexity
-			for (size_t pos = 0, prev = 0; pos < first_row.size(); ++pos) {
-				if (first_row[pos] == raw_csv->opts.delimiter ||
-					first_row[pos] == raw_csv->opts.terminator) {
-					raw_csv->col_names.emplace_back(first_row.data() + prev, pos - prev);
-					h_num_cols++;
-					if (first_row[pos] == raw_csv->opts.terminator) {
-						break;
-					}
-					prev = pos + 1;
-				}
-			}
-		} else {
-			h_num_cols = std::count(first_row.begin(), first_row.end(), raw_csv->opts.delimiter) + 1;
-			// assign column indexes as names if the header column is not present
-			raw_csv->col_names.reserve(h_num_cols);
-			const string str_prefix = args->prefix != nullptr ? string(args->prefix) : "";
-			for (int i = 0; i < h_num_cols; ++i) {
-				raw_csv->col_names.push_back(str_prefix + std::to_string(i));
-			}
+		error = setColumnNamesFromCsv(raw_csv);
+		if (error != GDF_SUCCESS) {
+			return error;
 		}
+		const int h_num_cols = raw_csv->col_names.size();
 
 		// Allocating a boolean array that will use to state if a column needs to read or filtered.
 		raw_csv->h_parseCol = (bool*)malloc(sizeof(bool) * (h_num_cols));
 		RMM_TRY( RMM_ALLOC((void**)&raw_csv->d_parseCol,(sizeof(bool) * (h_num_cols)),0 ) );
 		for (int i = 0; i<h_num_cols; i++)
 			raw_csv->h_parseCol[i]=true;
+		
+		// Rename empty column names to "Unnamed: col_index"
+		for (size_t col_idx = 0; col_idx < raw_csv->col_names.size(); ++col_idx) {
+			if (raw_csv->col_names[col_idx].empty()) {
+				raw_csv->col_names[col_idx] = string("Unnamed: ") + std::to_string(col_idx);
+			}
+		}
 
 		int h_dup_cols_removed = 0;
 		// Looking for duplicates
@@ -551,10 +619,7 @@ gdf_error read_csv(csv_read_arg *args)
 		CUDA_TRY(cudaMemcpy(raw_csv->d_parseCol, raw_csv->h_parseCol, sizeof(bool) * (raw_csv->num_actual_cols), cudaMemcpyHostToDevice));
 	}
 
-	if (raw_csv->header_row>=0) {
-		raw_csv->num_records-=1;
-	}
-	
+
 	//-----------------------------------------------------------------------------
 	//---  done with host data
 	if (args->input_data_form == gdf_csv_input_form::FILE_PATH)
@@ -912,7 +977,7 @@ gdf_error uploadDataToDevice(const char *h_uncomp_data, size_t h_uncomp_size,
 
   // Discard only blank lines, only fully comment lines, or both.
   // If only handling one of them, ensure it doesn't match against \0 as we do
-  // not certain scenarios filtered-out (end-of-file)
+  // not want certain scenarios to be filtered out (end-of-file)
   if (raw_csv->opts.skipblanklines || raw_csv->opts.comment != '\0') {
     const auto match1 = raw_csv->opts.skipblanklines ? raw_csv->opts.terminator
                                                      : raw_csv->opts.comment;
@@ -929,20 +994,20 @@ gdf_error uploadDataToDevice(const char *h_uncomp_data, size_t h_uncomp_size,
 
   raw_csv->num_records = h_rec_starts.size();
 
-  // Exclude the rows before the header row
-  if (raw_csv->header_row > 0) {
+  // Exclude the rows before the header row (inclusive)
+  // But copy the header data for parsing the column names later (if necessary)
+  if (raw_csv->header_row >= 0) {
+    raw_csv->header.assign(
+        h_uncomp_data + h_rec_starts[raw_csv->header_row],
+        h_uncomp_data + h_rec_starts[raw_csv->header_row + 1]);
     h_rec_starts.erase(h_rec_starts.begin(),
-                       h_rec_starts.begin() + raw_csv->header_row);
+                       h_rec_starts.begin() + raw_csv->header_row + 1);
     raw_csv->num_records = h_rec_starts.size();
   }
 
   // Exclude the rows that exceed past the requested number
   if (raw_csv->nrows >= 0 && raw_csv->nrows < raw_csv->num_records) {
-    if (raw_csv->header_row >= 0) {
-      h_rec_starts.resize(raw_csv->nrows + 2);  // include header & end offset
-    } else {
-      h_rec_starts.resize(raw_csv->nrows + 1);  // include end offset
-    }
+    h_rec_starts.resize(raw_csv->nrows + 1);    // include end offset
     raw_csv->num_records = h_rec_starts.size();
   }
 
@@ -1276,38 +1341,27 @@ __global__ void storeRecordStart(char *data, size_t chunk_offset,
  *
  * @return gdf_error GDF_SUCCESS upon completion
  *---------------------------------------------------------------------------**/
-gdf_error launch_dataConvertColumns(raw_csv_t *raw_csv, void **gdf, gdf_valid_type** valid, gdf_dtype* d_dtypes,string_pair **str_cols, unsigned long long *num_valid) {
+gdf_error launch_dataConvertColumns(raw_csv_t *raw_csv, void **gdf,
+                                    gdf_valid_type **valid, gdf_dtype *d_dtypes,
+                                    string_pair **str_cols,
+                                    unsigned long long *num_valid) {
+  int blockSize;    // suggested thread count to use
+  int minGridSize;  // minimum block count required
+  CUDA_TRY(cudaOccupancyMaxPotentialBlockSize(&minGridSize, &blockSize,
+                                              convertCsvToGdf));
 
-	int blockSize;		// suggested thread count to use
-	int minGridSize;	// minimum block count required
-	CUDA_TRY( cudaOccupancyMaxPotentialBlockSize(&minGridSize, &blockSize, convertCsvToGdf) );
+  // Calculate actual block count to use based on records count
+  int gridSize = (raw_csv->num_records + blockSize - 1) / blockSize;
 
-	// Calculate actual block count to use based on records count
-	int gridSize = (raw_csv->num_records + blockSize - 1) / blockSize;
+  convertCsvToGdf <<< gridSize, blockSize >>> (
+      raw_csv->data, raw_csv->opts, raw_csv->num_records,
+      raw_csv->num_actual_cols, raw_csv->d_parseCol, raw_csv->recStart,
+      d_dtypes,
+      raw_csv->d_naTrie.empty() ? nullptr : raw_csv->d_naTrie.data().get(), gdf,
+      valid, str_cols, num_valid);
 
-	auto first_data_rec_start = raw_csv->recStart;
-	if (raw_csv->header_row >= 0) {
-		// skip the header row if present
-		++first_data_rec_start;
-	}
-
-	convertCsvToGdf <<< gridSize, blockSize >>> (
-		raw_csv->data,
-		raw_csv->opts,
-		raw_csv->num_records,
-		raw_csv->num_actual_cols,
-		raw_csv->d_parseCol,
-		first_data_rec_start,
-		d_dtypes,
-		raw_csv->d_naTrie.empty() ? nullptr : raw_csv->d_naTrie.data().get(),
-		gdf,
-		valid,
-		str_cols,
-		num_valid
-	);
-
-	CUDA_TRY( cudaGetLastError() );
-	return GDF_SUCCESS;
+  CUDA_TRY(cudaGetLastError());
+  return GDF_SUCCESS;
 }
 
 /**---------------------------------------------------------------------------*
@@ -1353,6 +1407,53 @@ struct ConvertFunctor {
     value = convertStrToValue<T>(csvData, start, end, opts);
   }
 };
+
+/**---------------------------------------------------------------------------*
+ * @brief CUDA kernel iterates over the data until the end of the current field
+ * 
+ * Also iterates over (one or more) delimiter characters after the field.
+ *
+ * @param[in] raw_csv The entire CSV data to read
+ * @param[in] opts A set of parsing options
+ * @param[in] pos Offset to start the seeking from 
+ * @param[in] stop Offset of the end of the row
+ *
+ * @return long position of the last character in the field, including the 
+ *  delimiter(s) folloing the field data
+ *---------------------------------------------------------------------------**/
+__device__ 
+long seekFieldEnd(const char *raw_csv, const ParseOptions opts, long pos, long stop) {
+	bool quotation	= false;
+	while(true){
+		// Use simple logic to ignore control chars between any quote seq
+		// Handles nominal cases including doublequotes within quotes, but
+		// may not output exact failures as PANDAS for malformed fields
+		if(raw_csv[pos] == opts.quotechar){
+			quotation = !quotation;
+		}
+		else if(quotation==false){
+			if(raw_csv[pos] == opts.delimiter){
+				while (opts.multi_delimiter &&
+					   pos < stop &&
+					   raw_csv[pos + 1] == opts.delimiter) {
+					++pos;
+				}
+				break;
+			}
+			else if(raw_csv[pos] == opts.terminator){
+				break;
+			}
+			else if(raw_csv[pos] == '\r' && ((pos+1) < stop && raw_csv[pos+1] == '\n')){
+				stop--;
+				break;
+			}
+		}
+		if(pos>=stop)
+			break;
+		pos++;
+	}
+	return pos;
+}
 
 /**---------------------------------------------------------------------------*
  * @brief CUDA kernel that parses and converts CSV data into cuDF column data.
@@ -1401,36 +1502,13 @@ void convertCsvToGdf(char *raw_csv,
 	int  col 		= 0;
 	int  actual_col = 0;
 	int  stringCol 	= 0;
-	bool quotation	= false;
 
 	while(col<num_columns){
 
 		if(start>stop)
 			break;
 
-		while(true){
-			// Use simple logic to ignore control chars between any quote seq
-			// Handles nominal cases including doublequotes within quotes, but
-			// may not output exact failures as PANDAS for malformed fields
-			if(raw_csv[pos] == opts.quotechar){
-				quotation = !quotation;
-			}
-			else if(quotation==false){
-				if(raw_csv[pos] == opts.delimiter){
-					break;
-				}
-				else if(raw_csv[pos] == opts.terminator){
-					break;
-				}
-				else if(raw_csv[pos] == '\r' && ((pos+1) < stop && raw_csv[pos+1] == '\n')){
-					stop--;
-					break;
-				}
-			}
-			if(pos>=stop)
-				break;
-			pos++;
-		}
+		pos = seekFieldEnd(raw_csv, opts, pos, stop);
 
 		if(parseCol[col]==true){
 
@@ -1493,33 +1571,22 @@ void convertCsvToGdf(char *raw_csv,
  * @return gdf_error GDF_SUCCESS upon completion
  *---------------------------------------------------------------------------**/
 gdf_error launch_dataTypeDetection(raw_csv_t *raw_csv,
-                                   column_data_t *d_columnData)
-{
-	int blockSize;		// suggested thread count to use
-	int minGridSize;	// minimum block count required
-	CUDA_TRY( cudaOccupancyMaxPotentialBlockSize(&minGridSize, &blockSize, dataTypeDetection) );
+                                   column_data_t *d_columnData) {
+  int blockSize;    // suggested thread count to use
+  int minGridSize;  // minimum block count required
+  CUDA_TRY(cudaOccupancyMaxPotentialBlockSize(&minGridSize, &blockSize,
+                                              dataTypeDetection));
 
-	// Calculate actual block count to use based on records count
-	int gridSize = (raw_csv->num_records + blockSize - 1) / blockSize;
+  // Calculate actual block count to use based on records count
+  int gridSize = (raw_csv->num_records + blockSize - 1) / blockSize;
 
-	auto first_data_rec_start = raw_csv->recStart;
-	if (raw_csv->header_row >= 0) {
-		// skip the header row if present
-		++first_data_rec_start;
-	}
+  dataTypeDetection <<< gridSize, blockSize >>> (
+      raw_csv->data, raw_csv->opts, raw_csv->num_records,
+      raw_csv->num_actual_cols, raw_csv->d_parseCol, raw_csv->recStart,
+      d_columnData);
 
-	dataTypeDetection <<< gridSize, blockSize >>> (
-		raw_csv->data,
-		raw_csv->opts,
-		raw_csv->num_records,
-		raw_csv->num_actual_cols,
-		raw_csv->d_parseCol,
-		first_data_rec_start,
-		d_columnData
-	);
-
-	CUDA_TRY( cudaGetLastError() );
-	return GDF_SUCCESS;
+  CUDA_TRY(cudaGetLastError());
+  return GDF_SUCCESS;
 }
 
 /**---------------------------------------------------------------------------*
@@ -1560,7 +1627,6 @@ void dataTypeDetection(char *raw_csv,
 	long pos 		= start;
 	int  col 		= 0;
 	int  actual_col = 0;
-	bool quotation	= false;
 
 	// Going through all the columns of a given record
 	while(col<num_columns){
@@ -1568,31 +1634,7 @@ void dataTypeDetection(char *raw_csv,
 		if(start>stop)
 			break;
 
-		// Finding the breaking point for each column
-		while(true){
-			// Use simple logic to ignore control chars between any quote seq
-			// Handles nominal cases including doublequotes within quotes, but
-			// may not output exact failures as PANDAS for malformed fields
-			if(raw_csv[pos] == opts.quotechar){
-				quotation = !quotation;
-			}
-			else if(quotation==false){
-				if(raw_csv[pos] == opts.delimiter){
-					break;
-				}
-				else if(raw_csv[pos] == opts.terminator){
-					break;
-				}
-				else if(raw_csv[pos] == '\r' && ((pos+1) < stop && raw_csv[pos+1] == '\n')){
-					stop--;
-					break;
-				}
-			}
-			if(pos>=stop)
-				break;
-			pos++;
-		}
-
+		pos = seekFieldEnd(raw_csv, opts, pos, stop);
 
 		// Checking if this is a column that the user wants --- user can filter columns
 		if(parseCol[col]==true){
@@ -1702,50 +1744,3 @@ void dataTypeDetection(char *raw_csv,
 
 	}
 }
-
-//----------------------------------------------------------------------------------------------------------------
-
-/*
- * Return which bit is set
- * x is the occurrence: 1 = first, 2 = seconds, ...
- */
-__device__ int findSetBit(int tid, long num_bits, uint64_t *r_bits, int x) {
-
-	int idx = tid;
-
-	if ( x == 0 )
-		return -1;
-
-	int withinBitCount = 0;
-	int offset = 0;
-	int found  = 0;
-
-	uint64_t bitmap = r_bits[idx];
-
-	while (found != x)
-	{
-		if(bitmap == 0)
-		{
-			idx++;
-			if (idx >= num_bits)
-				return -1;
-			bitmap = r_bits[idx];
-			offset += 64;
-			withinBitCount = 0;
-		}
-
-		if ( bitmap & 1 ) {
-			found++;			//found a set bit
-		}
-
-		bitmap >>= 1;
-		++withinBitCount;
-	 }
-
-	offset += withinBitCount -1;
-
-
-	return offset;
-}
-
-