+++ cpp/tests/bitmask/valid_ops_test.cu
@@ -49,10 +49,12 @@ TEST_F(ValidsTest, NoValids)
 TEST_F(ValidsTest, NullValids)
 {
   int count{-1};
-  gdf_error error_code = gdf_count_nonzero_mask(nullptr, 1, &count);
+  const gdf_size_type size{100};
+  gdf_error error_code = gdf_count_nonzero_mask(nullptr, size, &count);
 
   ASSERT_EQ(GDF_SUCCESS,error_code) << "GDF Operation did not complete successfully.";
-  EXPECT_EQ(0, count);
+
+  EXPECT_EQ(size, count);
 }
 
 TEST_F(ValidsTest, NullCount)