+++ CHANGELOG.md
@@ -7,14 +7,19 @@
 - PR #807 CSV Reader: Add byte_range parameter to specify the range in the input file to be read
 - PR #858 Add series feature hashing support
 - PR #871 CSV Reader: Add support for NA values, including user specified strings
+- PR #893 Adds PyArrow based parquet readers / writers to Python, fix category dtype handling, fix arrow ingest buffer size issues
 - PR #867 CSV Reader: Add support for ignoring blank lines and comment lines
+- PR #887 Add Series digitize method
 - PR #895 Add Series groupby
 - PR #898 Add DataFrame.groupby(level=0) support
 - PR #888 CSV Reader: Add prefix parameter for column names, used when parsing without a header
+- PR #918 Add Series.groupby(level=0) support
+- PR #906 Add binary and comparison ops to DataFrame
 
 ## Improvements
 
 - PR #730 Improve performance of `gdf_table` constructor
+- PR #813 unified libcudf API functions by replacing gpu_ with gdf_
 - PR #822 Add support for `__cuda_array_interface__` for ingest
 - PR #756 Consolidate common helper functions from unordered map and multimap
 - PR #753 Improve performance of groupby sum and average, especially for cases with few groups.
@@ -24,6 +29,9 @@
 - PR #694 Unit test utilities improvements
 - PR #878 Add better indexing to Groupby
 - PR #554 Add `empty` method and `is_monotonic` attribute to `Index`
+- PR #909 CSV Reader: Avoid host->device->host copy for header row data
+- PR #916 Improved unit testing and error checking for `gdf_column_concat`
+- PR #943 Updated `count_nonzero_mask` to return `num_rows` when the mask is null
 
 
 ## Bug Fixes
@@ -40,6 +48,9 @@
 - PR #876 Added missing `<type_traits>` include
 - PR #889 Deleted test_rmm.py which has now moved to RMM repo
 - PR #866 Merge v0.5.1 numpy ABI hotfix into 0.6
+- PR #917 value_counts return int type on empty columns
+- PR #927 CSV Reader: Fix category GDF_CATEGORY hashes not being computed properly
+- PR #921 CSV Reader: Fix parsing errors with delim_whitespace, quotations in the header row, unnamed columns
 
 
 # cuDF 0.5.1 (05 Feb 2019)
@@ -60,6 +71,7 @@
 - PR #501 CSV Reader: Add support for user-specified decimal point and thousands separator to read_csv_strings()
 - PR #455 CSV Reader: Add support for user-specified decimal point and thousands separator to read_csv()
 - PR #439 add `DataFrame.drop` method similar to pandas
+- PR #356 add `DataFrame.transpose` method and `DataFrame.T` property similar to pandas
 - PR #505 CSV Reader: Add support for user-specified boolean values
 - PR #350 Implemented Series replace function
 - PR #490 Added print_env.sh script to gather relevant environment details when reporting cuDF issues